organize local database structure and create folders
