+++ app/src/main/java/edu/neu/madcourse/wellness_studio/Profile_With_Local_DB_Example.java
@@ -8,7 +8,7 @@ import androidx.appcompat.app.AppCompatActivity;
 import java.util.List;
 
 import localDatabase.AppDatabase;
-import localDatabase.User;
+import localDatabase.userInfo.User;
 
 public class Profile_With_Local_DB_Example extends AppCompatActivity {
     public TextView testTextView;
