+++ app/src/main/java/localDatabase/lightExercise/LightExerciseDao.java
@@ -1,4 +1,4 @@
-package localDatabase;
+package localDatabase.lightExercise;
 
 import androidx.room.Dao;
 import androidx.room.Delete;
@@ -7,8 +7,6 @@ import androidx.room.Query;
 import androidx.room.Update;
 
 import java.util.List;
-
-import localDatabase.LightExercise;
 
 @Dao
 public interface LightExerciseDao {
