+++ app/src/main/java/localDatabase/wakeUpAndSleepGoal/SleepGoal.java
@@ -1,4 +1,4 @@
-package localDatabase;
+package localDatabase.wakeUpAndSleepGoal;
 
 import androidx.annotation.NonNull;
 import androidx.room.ColumnInfo;
